+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 6.4.0 [unreleased]
 
 ### Features
+1. [#366](https://github.com/influxdata/influxdb-client-java/pull/356): Added an endpoint to query wit  InfluxQL (v1)
 1. [#373](https://github.com/influxdata/influxdb-client-java/pull/373):
    * Add ability to define imports for each flux function [FluxDSL]
    * Add ability use multiple flux expressions [FluxDSL]